9028 datamodellering reimplement support for converting between array and regular field
