+++ src/studio/src/designer/frontend/packages/schema-model/src/index.ts
@@ -1,7 +1,22 @@
-export * from './lib/build-ui-schema';
-export * from './lib/build-json-schema';
-export * from './lib/types';
-export * from './lib/restrictions';
-export * from './lib/selectors';
-export * from './lib/mutations';
-export * from './lib/utils';
+export { buildJsonSchema } from './lib/build-json-schema';
+export { buildUiSchema } from './lib/build-ui-schema';
+export { getRestrictions, StrRestrictionKeys } from './lib/restrictions';
+export type { UiSchemaNode, UiSchemaNodes } from './lib/types';
+export { ROOT_POINTER, ObjectKind, Keywords, FieldType, CombinationKind } from './lib/types';
+export { createNodeBase, replaceLastPointerSegment, makePointer } from './lib/utils';
+export { createChildNode } from './lib/mutations/create-node';
+export { removeNodeByPointer } from './lib/mutations/remove-node';
+export { renameNodePointer } from './lib/mutations/rename-node';
+export { promotePropertyToType } from './lib/mutations/promote-node';
+export { toggleArrayAndField, canToggleArrayAndField } from './lib/mutations/toggle-array-field';
+export {
+  combinationIsNullable,
+  getChildNodesByNode,
+  getChildNodesByPointer,
+  getNodeByPointer,
+  getNodeDisplayName,
+  getNodeIndexByPointer,
+  getRootNodes,
+  getUniqueNodePath,
+  pointerExists,
+} from './lib/selectors';
