+++ src/studio/src/designer/frontend/packages/schema-model/src/lib/mappers/field-type.ts
