+++ src/studio/src/designer/frontend/packages/schema-model/src/lib/mappers/generic.ts
